feat: fin indicateurs 2, 3
